feat(feedbacks): grant Reception role access and animate QR card - Add feedbacks.view / feedbacks.delete permissions, granted to   Super Admin, Admin, and Réception via seeder. - Gate backoffice feedback routes and UI actions with the new   permissions; menu link uses feedbacks.view. - Redesign QR page with an infinite rotating conic-gradient border,   pulsing glow, "Scannez-moi" chip and brand wordmark. Animations   are disabled cleanly when printing and for prefers-reduced-motion.

+++ routes/backoffice.php
@@ -25,6 +25,7 @@ use App\Http\Controllers\Backoffice\QuizController;
 use App\Http\Controllers\Backoffice\QuizQuestionController;
 use App\Http\Controllers\Backoffice\HelpController;
 use App\Http\Controllers\Backoffice\UserController;
+use App\Http\Controllers\Backoffice\EnseignantAccountController;
 use App\Http\Controllers\Backoffice\RoleController;
 use App\Http\Controllers\Backoffice\WhatsAppCampaignController;
 use App\Http\Controllers\Backoffice\NewsletterSubscriberController;
@@ -237,10 +238,10 @@ Route::prefix('backoffice')
         Route::prefix('feedbacks')
             ->name('feedbacks.')
             ->group(function () {
-                Route::get('/', [BackofficeFeedbackController::class, 'index'])->name('index');
-                Route::get('/qr', [BackofficeFeedbackController::class, 'qr'])->name('qr');
-                Route::get('/{id}', [BackofficeFeedbackController::class, 'show'])->where('id', '[0-9]+')->name('show');
-                Route::delete('/{id}', [BackofficeFeedbackController::class, 'destroy'])->where('id', '[0-9]+')->name('destroy');
+                Route::get('/', [BackofficeFeedbackController::class, 'index'])->middleware('permission:feedbacks.view')->name('index');
+                Route::get('/qr', [BackofficeFeedbackController::class, 'qr'])->middleware('permission:feedbacks.view')->name('qr');
+                Route::get('/{id}', [BackofficeFeedbackController::class, 'show'])->where('id', '[0-9]+')->middleware('permission:feedbacks.view')->name('show');
+                Route::delete('/{id}', [BackofficeFeedbackController::class, 'destroy'])->where('id', '[0-9]+')->middleware('permission:feedbacks.delete')->name('destroy');
             });
 
         /*
@@ -344,6 +345,22 @@ Route::prefix('backoffice')
                 Route::get('/{user}/edit', [UserController::class, 'edit'])->middleware('permission:users.edit')->name('edit');
                 Route::put('/{user}', [UserController::class, 'update'])->middleware('permission:users.edit')->name('update');
                 Route::delete('/{user}', [UserController::class, 'destroy'])->middleware('permission:users.delete')->name('destroy');
+            });
+
+        /*
+        |----------------------------------------------------------------------
+        | COMPTES ENSEIGNANTS (Login accounts for teachers)
+        |----------------------------------------------------------------------
+        */
+        Route::prefix('enseignant-accounts')
+            ->name('enseignant_accounts.')
+            ->group(function () {
+                Route::get('/', [EnseignantAccountController::class, 'index'])->middleware('permission:users.view')->name('index');
+                Route::get('/create', [EnseignantAccountController::class, 'create'])->middleware('permission:users.create')->name('create');
+                Route::post('/', [EnseignantAccountController::class, 'store'])->middleware('permission:users.create')->name('store');
+                Route::get('/{account}/edit', [EnseignantAccountController::class, 'edit'])->middleware('permission:users.edit')->name('edit');
+                Route::put('/{account}', [EnseignantAccountController::class, 'update'])->middleware('permission:users.edit')->name('update');
+                Route::delete('/{account}', [EnseignantAccountController::class, 'destroy'])->middleware('permission:users.delete')->name('destroy');
             });
 
         /*
